refactor(mutations): repo-list hooks grouping
#73
