+++ src/features/repo-list/index.tsx
@@ -2,75 +2,32 @@ import React from "react";
 import { Repo, Tabs, SimplePagination, Card } from "shared/components";
 import { str, dom } from "shared/helpers";
 import { useReposQuery } from "./queries.gen";
-import * as Params from "./params";
-import useStarred from "./hooks";
+import { useFilters, PAGE_SIZE, useStarring } from "./hooks";
+import { tabsMap } from "./params";
 import "./index.scss";
 
 type Props = {
     username: string;
 };
 
-const PAGE_SIZE = 30;
-
 /**
- * @hook Работа с фильтрацией по affilations, пагинацией
+ * @feature Список репозиториев пользователя
+ * FIXME: rename to UserRepoList? (coz - user as dep)
  */
-const useFilters = () => {
-    const { tab, setTab, tabEnum } = Params.useTabParam();
-    const { after, before, setCursor } = Params.useCursorParam();
-
-    /**
-     * Обработчик выбора вкладки
-     * @remark Реактивно сбрасываем пагинацию, при смене вкладки
-     */
-    const handleTabClick: typeof setTab = (type) => {
-        setTab(type);
-        setCursor({});
-    };
-
-    /**
-     * Обработчик пагинации
-     * @remark Явно определяем вкладку, чтобы она точно была задана (для соответствия ссылок features/origin)
-     */
-    const handlePaginationClick: typeof setCursor = (pageInfo) => {
-        setTab(tab);
-        setCursor(pageInfo);
-    };
-
-    return {
-        config: {
-            tab,
-            ownerAffiliations: [tabEnum],
-            after,
-            before,
-            /**
-             * @variant (!before, !after) => Первый вход, фетчим первые {PAGE_SIZE}
-             * @variant (after, !before) => След. страница, фетчим след. первые {PAGE_SIZE}
-             * @variant (!after, before) => Пред. страница, фетчим пред. последние {PAGE_SIZE}
-             * @variant (after, before) => (невозможна из-за реализации)
-             */
-            first: (!before && PAGE_SIZE) || undefined,
-            last: (before && PAGE_SIZE) || undefined,
-        },
-        handleTabClick,
-        handlePaginationClick,
-    };
-};
-// FIXME: rename to UserRepoList? (coz - user as dep)
 const RepoList = ({ username }: Props) => {
     const { handleTabClick, handlePaginationClick, config } = useFilters();
     const { data, loading, variables } = useReposQuery({
         variables: { login: username, ...config },
     });
     // TODO: transmit id and viewerHasStarred of nodes to handler func
-    const { handleStarring } = useStarred(variables);
+    const { handleStarring } = useStarring(variables);
     const { pageInfo, totalCount = 0, nodes } = data?.user?.repositories || {};
     const length = nodes?.length;
 
     return (
         <div className="repo-list">
             <Tabs className="repo-list__tabs">
-                {Object.keys(Params.tabsMap).map((type) => (
+                {Object.keys(tabsMap).map((type) => (
                     <Tabs.Item
                         key={type}
                         name={str.capitalize(type)}
@@ -88,7 +45,7 @@ const RepoList = ({ username }: Props) => {
                 {length !== 0 ? (
                     data?.user?.repositories.nodes?.map((node) => (
                         <Repo
-                            onClick={() => handleStarring(node?.id, node?.viewerHasStarred)}
+                            onClick={() => handleStarring(null, node?.viewerHasStarred)}
                             key={node?.id}
                             {...node}
                         />
